requirements.txt
PyQt5
requests
google-generativeai